requirements-dev.txt
# 开发和测试依赖
# Development and Testing Dependencies

# 测试框架 Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-html==4.1.1
pytest-xdist>=3.3.0
pytest-mock>=3.12.0
pytest-timeout==2.2.0

# 代码质量 Code Quality
black>=23.11.0
isort>=5.12.0
flake8==6.1.0
mypy==1.7.1
radon==6.0.1

# 安全检查 Security
safety>=3.0.0
bandit==1.7.5
pip-audit==2.6.1

# 性能测试 Performance Testing
locust==2.17.0

# 开发工具 Development Tools
pre-commit==3.5.0
watchdog==3.0.0
ipython==8.17.2
jupyter==1.0.0

# 文档生成 Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# 调试工具 Debugging
pdbpp==0.10.3
icecream==2.1.3

# 类型检查 Type Checking
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14
types-PyYAML==6.0.12.12

# 依赖管理 Dependency Management
pip-tools==7.3.0

# 环境管理 Environment Management
python-dotenv==1.0.0

# HTTP客户端测试 HTTP Client Testing
httpx==0.25.2
requests-mock==1.11.0

# 数据库测试 Database Testing
sqlalchemy-utils==0.41.1
factory-boy==3.3.0
faker==20.1.0

# 时间处理测试 Time Handling Testing
freezegun==1.2.2

# 配置文件处理 Configuration File Handling
toml==0.10.2

# 日志测试 Logging Testing
loguru==0.7.2

# 内存分析 Memory Profiling
memory-profiler==0.61.0
psutil==5.9.6

# 代码覆盖率可视化 Coverage Visualization
coverage[toml]==7.3.2